boilerplate for generic extended object API
